Refactor code and seperate logic into modules.
    - utility module split into three sepearete modules.
